RED-5108 - Fixing import error
